requirements.txt
flask==2.3.3
flask-bootstrap-components==0.2.0
flask-wtf==1.2.1
bootstrap-flask==2.2.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.8
